BW/8 - Refactoring of some methods to allow for dealing with the inbound request type and http method.
Handled the response codes, but i'm not totally excited by the hacky solution I up with, but its functional.  Project for a different day to find a better solution. Handling 404 a little differently. Likely better and more elegant solutions, but this will do for now.

As it stands, I haven't punished this thing yet, but its standing up to a lot of requests.  I'll have to obsess over it more tomorrow.